<Refactor> 엔진 리펙토링 (RenderSystem 책임 분리, DescriptorAllocator 추가)

+++ MarchingCubes/Core/Geometry/MarchingCubes/CPU/MC33/MC33TerrainBackend.cpp
@@ -22,13 +22,13 @@ void MC33TerrainBackend::requestRemesh(const RemeshRequest& req)
         grd->N[1] = chunkSize;
         grd->N[2] = chunkSize;
 
-        grd->d[0] = m_gridDesc.cellsize;
-        grd->d[1] = m_gridDesc.cellsize;
-        grd->d[2] = m_gridDesc.cellsize;
+        grd->d[0] = static_cast<double>(m_gridDesc.cellsize);
+        grd->d[1] = static_cast<double>(m_gridDesc.cellsize);
+        grd->d[2] = static_cast<double>(m_gridDesc.cellsize);
 
-        grd->r0[0] = m_gridDesc.origin.x + static_cast<float>(baseX) * m_gridDesc.cellsize;
-        grd->r0[1] = m_gridDesc.origin.y + static_cast<float>(baseY) * m_gridDesc.cellsize;
-        grd->r0[2] = m_gridDesc.origin.z + static_cast<float>(baseZ) * m_gridDesc.cellsize;
+        grd->r0[0] = static_cast<double>(m_gridDesc.origin.x + static_cast<float>(baseX) * m_gridDesc.cellsize);
+        grd->r0[1] = static_cast<double>(m_gridDesc.origin.y + static_cast<float>(baseY) * m_gridDesc.cellsize);
+        grd->r0[2] = static_cast<double>(m_gridDesc.origin.z + static_cast<float>(baseZ) * m_gridDesc.cellsize);
 
         grd->nonortho = 0;
         grd->periodic = 0;
